lib/deploy: move some file system helpers to libotutil
The copy_dir_recurse(), dirfd_copy_attributes_and_xattrs() and is_ro_mount()
functions could be used in other places, so move them as libotutil helpers.
