+++ src/libostree/ostree-sysroot-private.h
@@ -28,19 +28,6 @@
 
 G_BEGIN_DECLS
 
-typedef enum {
-
-  /* Don't flag deployments as immutable. */
-  OSTREE_SYSROOT_DEBUG_MUTABLE_DEPLOYMENTS = 1 << 0,
-  /* See https://github.com/ostreedev/ostree/pull/759 */
-  OSTREE_SYSROOT_DEBUG_NO_XATTRS = 1 << 1,
-  /* https://github.com/ostreedev/ostree/pull/1049 */
-  OSTREE_SYSROOT_DEBUG_TEST_FIFREEZE = 1 << 2,
-  /* This is a temporary flag until we fully drop the explicit `systemctl start
-   * ostree-finalize-staged.service` so that tests can exercise the new path unit. */
-  OSTREE_SYSROOT_DEBUG_TEST_STAGED_PATH = 1 << 3,
-} OstreeSysrootDebugFlags;
-
 /**
  * OstreeSysroot:
  * Internal struct
